// packages/query/types/inscription.d.ts
export interface InscriptionResponseHiro {
    address: string;
    content_length: number;
    content_type: string;
    curse_type: string | null;
    genesis_address: string;
    genesis_block_hash: string;
    genesis_block_height: number;
    genesis_fee: string;
    genesis_timestamp: number;
    genesis_tx_id: string;
    id: string;
    location: string;
    mime_type: string;
    number: number;
    offset: string;
    output: string;
    recursive: boolean;
    recursion_refs: string | null;
    sat_coinbase_height: number;
    sat_ordinal: string;
    sat_rarity: string;
    timestamp: number;
    tx_id: string;
    value: string;
}
//# sourceMappingURL=inscription.d.ts.map